requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.5
colorclass==2.2.2
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==1.1.2
Django==3.2.13
django-allauth==0.51.0
django-cas-ng==4.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
lxml==4.9.1
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
pep517==0.13.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-cas==1.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0